Add codesandbox project creation support
