+++ packages/docs-infra/src/lite/runtime/index.ts
@@ -7,6 +7,17 @@ export {
 } from './abstractCreateDemo';
 export { CodeHighlighter } from './CodeHighlighter';
 export {
+  createCodeSandbox,
+  openCodeSandbox,
+  type CodeSandboxProject,
+  type CreateCodeSandboxOptions,
+} from './createCodeSandbox';
+export {
+  createSandboxFileSystem,
+  type CreateSandboxFileSystemOptions,
+  type SandboxFileSystem,
+} from './createSandboxFileSystem';
+export {
   createStackBlitz,
   openStackBlitz,
   type CreateStackBlitzOptions,
